# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for AutoBan.",
        "title": "AutoBan API",
        "contact": {},
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/vehicles/{vehicle_id}/oil-changes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all oil changes for a user vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Changes"
                ],
                "summary": "List all oil changes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListOilChangesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/oil-changes/last": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the last oil change for a user vehicle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Changes"
                ],
                "summary": "Last oil change",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OilChangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/oil-changes/{oil_change_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get information about a specific oil change",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Changes"
                ],
                "summary": "Get oil change information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Oil change ID",
                        "name": "oil_change_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OilChangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/oil-filters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all oil filter change records for a specific user vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Filters"
                ],
                "summary": "List oil filter changes for a user vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListOilFiltersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/oil-filters/last": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the most recent oil filter change record for a specific user vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Filters"
                ],
                "summary": "Get last oil filter change for a user vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OilFilterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/user/vehicles/{vehicle_id}/oil-filters/{oil_filter_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific oil filter change record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Oil Filters"
                ],
                "summary": "Get oil filter change by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Oil filter change ID",
                        "name": "oil_filter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OilFilterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ListOilChangesResponse": {
            "description": "Oil change list response",
            "type": "object",
            "properties": {
                "oil_changes": {
                    "description": "Oil changes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OilChangeResponse"
                    }
                }
            }
        },
        "dto.ListOilFiltersResponse": {
            "description": "List of oil filter changes",
            "type": "object",
            "properties": {
                "oil_filters": {
                    "description": "List of oil filter changes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OilFilterResponse"
                    }
                }
            }
        },
...
        "dto.OilChangeResponse": {
            "description": "Oil change response",
            "type": "object",
            "properties": {
                "change_date": {
                    "description": "Change date",
                    "type": "string",
                    "example": "2021-01-01"
                },
                "change_mileage": {
                    "description": "Change mileage",
                    "type": "integer",
                    "example": 10000
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "next_change_date": {
                    "description": "Next change date",
                    "type": "string",
                    "example": "2021-01-01"
                },
                "next_change_mileage": {
                    "description": "Next change mileage",
                    "type": "integer",
                    "example": 20000
                },
                "notes": {
                    "description": "Notes",
                    "type": "string",
                    "example": "تعویض روغن"
                },
                "oil_brand": {
                    "description": "Oil brand",
                    "type": "string",
                    "example": "بهران"
                },
                "oil_capacity": {
                    "description": "Oil capacity",
                    "type": "number",
                    "example": 5
                },
                "oil_name": {
                    "description": "Oil name",
                    "type": "string",
                    "example": "تکتاز"
                },
                "oil_type": {
                    "description": "Oil type",
                    "type": "string",
                    "example": "مینرال، سنتتیک، نیمه سنتتیک"
                },
                "oil_viscosity": {
                    "description": "Oil viscosity",
                    "type": "string",
                    "example": "5W-30, 10W-40, etc."
                },
                "service_center": {
                    "description": "Service center",
                    "type": "string",
                    "example": "اتوبان سرویس"
                },
                "user_vehicle_id": {
                    "description": "User vehicle ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.OilFilterResponse": {
            "description": "Oil filter change response",
            "type": "object",
            "properties": {
                "change_date": {
                    "description": "Date when filter was changed",
                    "type": "string"
                },
                "change_mileage": {
                    "description": "Mileage when filter was changed",
                    "type": "integer"
                },
                "filter_brand": {
                    "description": "Brand of the oil filter",
                    "type": "string"
                },
                "filter_name": {
                    "description": "Name of the oil filter",
                    "type": "string"
                },
                "filter_part_number": {
                    "description": "Part number of the oil filter",
                    "type": "string"
                },
                "filter_type": {
                    "description": "Type of the oil filter",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the oil filter change",
                    "type": "integer"
                },
                "next_change_date": {
                    "description": "Next change date",
                    "type": "string"
                },
                "next_change_mileage": {
                    "description": "Next change mileage",
                    "type": "integer"
                },
                "notes": {
                    "description": "Additional notes",
                    "type": "string"
                },
                "service_center": {
                    "description": "Service center where filter was changed",
                    "type": "string"
                },
                "user_vehicle_id": {
                    "description": "ID of the user vehicle",
                    "type": "integer"
                }
            }
        },
...
        "errors.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable code (stable across languages)",
                    "type": "string"
                },
                "details": {
                    "description": "Optional structured details for debugging (safe to expose)",
                    "type": "object",
                    "additionalProperties": {}
                },
                "fields": {
                    "description": "Optional list of field-level errors (validation)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errors.FieldError"
                    }
                },
                "message": {
                    "description": "Bilingual message for human-readable display",
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorMessage"
                        }
                    ]
                },
                "retryable": {
                    "description": "Whether the operation can be retried by client later",
                    "type": "boolean"
                }
            }
        },
        "errors.ErrorMessage": {
            "type": "object",
            "properties": {
                "english": {
                    "type": "string"
                },
                "persian": {
                    "type": "string"
                }
            }
        },
        "errors.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/errors.ErrorMessage"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}